+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.7.0 (unreleased)
 
 - Added check for `add_foreign_key` with MySQL and MariaDB
+- Added check for `add_column` with callable default value with MySQL and MariaDB
 
 ## 2.6.0 (2026-04-07)